refactor(solver-wasm): Unite SimulationEnvironment
