+++ packages/solver-wasm/src/lib.rs
@@ -8,6 +8,7 @@ pub mod puyo;
 pub mod puyo_attr;
 pub mod puyo_coord;
 pub mod puyo_type;
+pub mod simulation_environment;
 pub mod simulator;
 pub mod simulator_bb;
 pub mod solution;
@@ -22,7 +23,8 @@ extern crate num_derive;
 use exploration_target::ExplorationTarget;
 use puyo::{Field, NextPuyos};
 use puyo_coord::PuyoCoord;
-use simulator::{SimulationEnvironment, Simulator};
+use simulation_environment::SimulationEnvironment;
+use simulator::Simulator;
 use solution_explorer::SolutionExplorer;
 use std::collections::HashSet;
 use wasm_bindgen::prelude::*;
