+++ packages/solver-wasm/src/simulator.rs
@@ -13,6 +13,7 @@ use crate::{
     puyo_attr::{PuyoAttr, SPECIAL_ATTRS},
     puyo_coord::PuyoCoord,
     puyo_type::{convert_type, get_attr, is_chance_type, is_colored_type, is_plus_type, PuyoType},
+    simulation_environment::SimulationEnvironment,
     trace_mode::*,
 };
 
@@ -23,16 +24,6 @@ type Block = HashMap<PuyoCoord, Puyo>;
 pub struct BlockWithAttr {
     attr: PuyoAttr,
     block: Block,
-}
-
-#[derive(Debug, Serialize, Deserialize)]
-pub struct SimulationEnvironment {
-    pub is_chance_mode: bool,
-    pub minimum_puyo_num_for_popping: u32,
-    pub max_trace_num: u32,
-    pub trace_mode: TraceMode,
-    pub popping_leverage: f64,
-    pub chain_leverage: f64,
 }
 
 #[derive(Debug)]
